requirements.txt
yt-dlp
requests
pillow